+++ src/Resources/Customer.php
@@ -319,7 +319,7 @@ class Customer extends AbstractUnzerResource
     /**
      * {@inheritDoc}
      */
-    protected function getResourcePath(): string
+    protected function getResourcePath($httpMethod = HttpAdapterInterface::REQUEST_GET): string
     {
         return 'customers';
     }