+++ src/Resources/Webhook.php
@@ -91,7 +91,7 @@ class Webhook extends AbstractUnzerResource
     /**
      * {@inheritDoc}
      */
-    protected function getResourcePath(): string
+    protected function getResourcePath($httpMethod = HttpAdapterInterface::REQUEST_GET): string
     {
         return 'webhooks';
     }